requirements.txt
Pillow==9.4.0
requests==2.28.2
selenium==4.7.2
webdriver-manager==3.8.5